+++ activiti-spring-boot/spring-boot-samples/spring-boot-sample-hal-rest-api/src/test/java/org/activiti/TestConfiguration.java
@@ -43,7 +43,8 @@ public class TestConfiguration {
         converter.setObjectMapper(mapper);
 
         return new RestTemplateBuilder()
-                .additionalMessageConverters(converter);
+                .additionalMessageConverters(converter)
+                .additionalCustomizers();
     }
 
 }
